+++ CGSSGuide/AppDelegate.swift
@@ -38,7 +38,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
         
         // 设置SDWebImage过期时间
-        SDImageCache.shared().maxCacheAge = 60 * 60 * 24 * 365 * 10 // 10年缓存时间
+        SDImageCache.shared().config.maxCacheAge = 60 * 60 * 24 * 365 * 10 // 10年缓存时间
         
         // 初始化DrawerController和TabBarController
         let sb = UIStoryboard.init(name: "Main", bundle: nil)