+++ CGSSGuide/Card/Controller/Card3DModelController.swift
@@ -174,11 +174,11 @@ class Card3DModelController: BaseViewController {
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        navigationController?.setToolbarHidden(false, animated: true)
+        // navigationController?.setToolbarHidden(false, animated: true)
     }
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
-        navigationController?.setToolbarHidden(true, animated: true)
+        // navigationController?.setToolbarHidden(true, animated: true)
     }