+++ CGSSGuide/Settings/Controller/WipeTableViewCell.swift
@@ -29,6 +29,7 @@ class WipeTableViewCell: UITableViewCell {
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+    
 
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)